requirements.txt
python-dotenv==1.0.1
pyTelegramBotAPI==4.16.1
openai==1.42.0
pymilvus==2.4.5
sentence-transformers==3.0.1
langchain==0.2.13
langchain-core==0.2.30
langdetect==1.0.9
langsmith==0.1.75
langchain-community==0.2.12
FlagEmbedding==1.2.11
peft==0.12.0
rank-bm25==0.2.2
psycopg2-binary==2.9.9
